+++ src/revision/CHANGE_LOG.md
@@ -5,6 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project/module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## V1.2.0 - xx.xx.2022
+
+### Added
+ - Implementation of PWM timer low level driver
+
+### Changed
+ - Remap LED low level drivers from GPIO to PWM timer
+
+### Memory usage:
+ - RAM: xkB/256kB (x%)
+ - FLASH: xkB/1024kB (x%)
+
+---
 ## V1.1.0 - 21.11.2022
 
 ### Added